requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.26.post2
Django==1.11
django-celery==3.2.2
django-filter==2.0.0
djangorestframework==3.8.2
kombu==3.0.37
Markdown==3.0.1
pytz==2018.6
vine==1.1.4
